+++ org/xbill/DNS/KXRecord.java
@@ -21,7 +21,7 @@ private
 KXRecord() {}
 
 private
-KXRecord(Name name, short dclass, int ttl) {
+KXRecord(Name name, int dclass, int ttl) {
 	super(name, Type.KX, dclass, ttl);
 }
 
@@ -37,12 +37,12 @@ getMember() {
  * @param target The host that authority is delegated to
  */
 public
-KXRecord(Name name, short dclass, int ttl, int preference, Name target) {
+KXRecord(Name name, int dclass, int ttl, int preference, Name target) {
 	super(name, Type.KX, dclass, ttl, preference, target);
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -50,7 +50,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	return rdataFromString(new KXRecord(name, dclass, ttl), st, origin);
