+++ org/xbill/DNS/MXRecord.java
@@ -21,7 +21,7 @@ private
 MXRecord() {}
 
 private
-MXRecord(Name name, short dclass, int ttl) {
+MXRecord(Name name, int dclass, int ttl) {
 	super(name, Type.MX, dclass, ttl);
 }
 
@@ -37,13 +37,13 @@ getMember() {
  * @param target The host that mail is sent to
  */
 public
-MXRecord(Name name, short dclass, int ttl, int priority, Name target)
+MXRecord(Name name, int dclass, int ttl, int priority, Name target)
 {
 	super(name, Type.MX, dclass, ttl, priority, target);
 }
 
 Record
-rrFromWire(Name name, short type, short dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, int ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -51,7 +51,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, short dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	return rdataFromString(new MXRecord(name, dclass, ttl), st, origin);
